# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ERP EDU — API de gestión educativa
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students/{id}/deploy:
    post:
      operationId: deployStudent
      tags: [students]
      summary: Desplegar Dolibarr propio del alumno (crear BD + contenedor)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resultado del despliegue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeployResult"
        "404":
          description: Alumno no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Docker o MariaDB no disponible / BASE_DOMAIN sin configurar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /students/{id}/dolibarr/state:
    get:
      operationId: getStudentContainerState
      tags: [students]
      summary: Estado del contenedor Dolibarr del alumno
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Estado actual del contenedor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DolibarrContainerState" }
        "404":
          description: Alumno no encontrado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /students/{id}/dolibarr/start:
    post:
      operationId: startStudentContainer
      tags: [students]
      summary: Arrancar el contenedor Dolibarr del alumno
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Contenedor arrancado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DolibarrContainerState" }
        "404":
          description: Alumno o contenedor no encontrado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "503":
          description: Docker no disponible
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /students/{id}/dolibarr/stop:
    post:
      operationId: stopStudentContainer
      tags: [students]
      summary: Detener el contenedor Dolibarr del alumno
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Contenedor detenido
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DolibarrContainerState" }
        "404":
          description: Alumno o contenedor no encontrado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "503":
          description: Docker no disponible
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /students/{id}/dolibarr:
    delete:
      operationId: destroyStudentContainer
      tags: [students]
      summary: Eliminar contenedor y BD del Dolibarr del alumno (datos perdidos)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Contenedor y BD eliminados
        "404":
          description: Alumno no encontrado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "503":
          description: Docker no disponible
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /groups/{id}/deploy-all:
    post:
      operationId: deployGroupStudents
      tags: [groups]
      summary: Desplegar todos los alumnos pendientes de un grupo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resultado del despliegue masivo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeployAllResult"
        "404":
          description: Grupo no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Employees (nóminas) ───────────────────────────────
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: Listar trabajadores
      parameters:
        - in: query
          name: studentId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de trabajadores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Crear trabajador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Trabajador creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Obtener trabajador
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trabajador
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Actualizar trabajador
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "200":
          description: Trabajador actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Eliminar trabajador
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Eliminado

  # ── Payrolls (nóminas) ────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Teacher ──────────────────────────────────────────
...
    DeployResult:
      type: object
      required: [studentId, status]
      properties:
        studentId:        { type: integer }
        status:           { type: string, enum: [synced, error, skipped] }
        containerName:    { type: ["string", "null"], description: "Nombre del contenedor Docker creado para el alumno." }
        publicUrl:        { type: ["string", "null"], description: "URL pública del Dolibarr del alumno (https://{user}.{baseDomain}/)." }
        containerState:   { type: ["string", "null"], description: "Estado del contenedor: running, exited, created, etc." }
        dolibarrPassword: { type: ["string", "null"] }
        error:            { type: ["string", "null"] }
    DolibarrContainerState:
      type: object
      required: [studentId, exists, state, publicUrl, containerName]
      properties:
        studentId:     { type: integer }
        exists:        { type: boolean }
        state:         { type: string, description: "running | exited | created | restarting | paused | dead | absent" }
        publicUrl:     { type: ["string", "null"] }
        containerName: { type: string }
        startedAt:     { type: ["string", "null"], format: date-time }
    DeployAllResult:
      type: object
      required: [deployed, skipped, errors]
      properties:
        deployed: { type: integer }
        skipped:  { type: integer }
        errors:
          type: array
          items:
            type: object
            required: [studentId, username, error]
            properties:
              studentId: { type: integer }
              username:  { type: string }
              error:     { type: string }

    # ── Auth ─────────────────────────────────────────────
...
    Employee:
      type: object
      required: [id, studentId, firstName, lastName, jobTitle, contractType, groupCategory, salaryBase, extraPayments, irpfRate, active, dolibarrSyncStatus, createdAt]
      properties:
        id:                   { type: integer }
        studentId:            { type: integer }
        firstName:            { type: string }
        lastName:             { type: string }
        dni:                  { type: ["string", "null"] }
        jobTitle:             { type: string }
        contractType:         { type: string, enum: [indefinido, temporal] }
        groupCategory:        { type: integer }
        salaryBase:           { type: number }
        extraPayments:        { type: integer, enum: [12, 14] }
        irpfRate:             { type: number }
        active:               { type: boolean }
        dolibarrEmployeeId:   { type: ["integer", "null"] }
        dolibarrSyncStatus:   { type: string, enum: [pending, synced, error] }
        dolibarrSyncError:    { type: ["string", "null"] }
        createdAt:            { type: string, format: date-time }
    EmployeeInput:
      type: object
      required: [studentId, firstName, lastName, jobTitle, salaryBase]
      properties:
        studentId:     { type: integer }
        firstName:     { type: string, minLength: 1 }
        lastName:      { type: string, minLength: 1 }
        dni:           { type: string }
        jobTitle:      { type: string, minLength: 1 }
        contractType:  { type: string, enum: [indefinido, temporal], default: indefinido }
        groupCategory: { type: integer, minimum: 1, maximum: 11, default: 7 }
        salaryBase:    { type: number, minimum: 0 }
        extraPayments: { type: integer, enum: [12, 14], default: 14 }
        irpfRate:      { type: number, minimum: 0, maximum: 45, default: 15 }

    # ── Payroll ───────────────────────────────────────────
...